+++ Core/JContent.cs
@@ -41,7 +41,7 @@ namespace Greatbone.Core
             level--;
         }
 
-        public void PutArr<T>(T[] v, ushort x = 0) where T : IPersist
+        public void PutArr<T>(T[] v, uint x = 0) where T : IPersist
         {
             PutArr(delegate
             {
@@ -65,7 +65,7 @@ namespace Greatbone.Core
             level--;
         }
 
-        public void PutObj<T>(T v, ushort x = 0) where T : IPersist
+        public void PutObj<T>(T v, uint x = 0) where T : IPersist
         {
             PutObj(delegate
             {
@@ -331,7 +331,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public JContent Put<V>(string name, V v, ushort x = 0) where V : IPersist
+        public JContent Put<V>(string name, V v, uint x = 0) where V : IPersist
         {
             if (counts[level]++ > 0) Add(',');
 
@@ -532,7 +532,7 @@ namespace Greatbone.Core
         }
 
 
-        public JContent Put<V>(string name, V[] v, ushort x = 0) where V : IPersist
+        public JContent Put<V>(string name, V[] v, uint x = 0) where V : IPersist
         {
             if (counts[level]++ > 0) Add(',');
